Decompose session instance file
