+++ scripts/file-length-allowlist.txt
@@ -17,7 +17,6 @@
 # be split, not grandfathered.
 
 # --- production code (limit 1000) ---
-session/instance.go 1400
 ui/sidebar.go 1258
 config/config.go 1128
 session/backend_hook.go 1026
