+++ session/instance.go
@@ -1,11 +1,8 @@
 package session
 
 import (
-	"crypto/rand"
 	"fmt"
 	"os"
-	"path/filepath"
-	"strings"
 	"sync"
 	"time"
 
@@ -553,266 +550,6 @@ func (i *Instance) setTmuxLocked(ts *tmux.TmuxSession) {
 	i.Tabs[0].tmux = ts
 }
 
-// ToInstanceData converts an Instance to its serializable form
-func (i *Instance) ToInstanceData() InstanceData {
-	i.mu.RLock()
-	defer i.mu.RUnlock()
-
-	data := InstanceData{
-		ID:     i.ID,
-		Title:  i.Title,
-		Path:   i.Path,
-		Branch: i.Branch,
-		// Persist BOTH axes: Liveness is the new canonical value, Status the
-		// legacy int kept for one release so a binary rolled back to before
-		// #1195 still reads a sensible status (and as the read-fallback source
-		// for records this build wrote before the split). SaveInstances skips
-		// transient (Loading/Deleting) rows, so a persisted Status is always a
-		// settled liveness value.
-		Status:     i.statusLocked(),
-		Liveness:   i.liveness,
-		Height:     i.Height,
-		Width:      i.Width,
-		CreatedAt:  i.CreatedAt,
-		UpdatedAt:  time.Now(),
-		Program:    i.Program,
-		AutoYes:    i.AutoYes,
-		Prompt:     i.Prompt,
-		UserKilled: i.userKilled,
-	}
-
-	if i.backend != nil {
-		data.BackendType = i.backend.Type()
-	}
-	if i.remoteMeta != nil {
-		data.RemoteMeta = i.remoteMeta
-	}
-
-	// Persist the usage-limit reset time only while the session is actually
-	// limit-blocked (#1146). Gating on the liveness keeps a recovered session
-	// from carrying a stale reset time to disk or into the snapshot — the
-	// in-memory field lingers after ClearLimitReached but is never serialized.
-	if i.liveness == LiveLimitReached {
-		data.LimitResetAt = i.limitResetAt
-	}
-
-	// Persist each tab so the full agent+shell tab list survives a restart
-	// (Sachin's hard requirement for #930): on reload FromInstanceData restores
-	// each local tab's tmux session by its exact persisted name, reconnecting
-	// live sessions across an af/daemon restart. Remote tabs (agent + optional
-	// terminal) carry no tmux session, so they serialize with an empty TmuxName;
-	// on restore HookBackend.Start re-derives them from the live terminal_cmd
-	// config (syncRemoteTabs) rather than from this serialized list, so a
-	// terminal_cmd added or removed while af was down is honored.
-	for _, tab := range i.Tabs {
-		td := TabData{Name: tab.Name, Kind: tab.Kind, Command: tab.Command}
-		if tab.tmux != nil {
-			td.TmuxName = tab.tmux.SanitizedName()
-		}
-		td.Conversation = conversationDataPtr(tab.Conversation)
-		data.Tabs = append(data.Tabs, td)
-	}
-
-	// Keep writing the legacy single TmuxName field (set from the agent tab) for
-	// one release: a binary rolled back to before #930 PR 2 still finds the
-	// agent session by its exact name, and old readers ignore the new Tabs list.
-	if ts := i.tmuxLocked(); ts != nil {
-		data.TmuxName = ts.SanitizedName()
-	}
-	if len(i.Tabs) > 0 {
-		data.AgentConversation = conversationDataPtr(i.Tabs[0].Conversation)
-	}
-
-	// Only include worktree data if gitWorktree is initialized
-	if i.gitWorktree != nil {
-		branchCreatedByUs := i.gitWorktree.BranchCreatedByUs()
-		// ExternalWorktree is true for in-place sessions (`af sessions create
-		// --here`, which attach to the repo's own working tree) and for
-		// instances persisted by the pre-#930-PR-3 create-on-existing-worktree
-		// feature. Cleanup() honors it by skipping removal of the user-owned
-		// worktree+branch. (BranchCreatedByUs is independent — it also flips
-		// false on the normal path when Setup reuses an existing branch; see
-		// git/worktree_ops.go setupFromExistingBranch.)
-		data.Worktree = GitWorktreeData{
-			RepoPath:          i.gitWorktree.GetRepoPath(),
-			WorktreePath:      i.gitWorktree.GetWorktreePath(),
-			SessionName:       i.Title,
-			BranchName:        i.gitWorktree.GetBranchName(),
-			BaseCommitSHA:     i.gitWorktree.GetBaseCommitSHA(),
-			ExternalWorktree:  i.gitWorktree.IsExternalWorktree(),
-			BranchCreatedByUs: &branchCreatedByUs,
-		}
-	}
-
-	// Only include PR info if it exists
-	if i.prInfo != nil {
-		data.PRInfo = PRInfoData{
-			Number: i.prInfo.Number,
-			Title:  i.prInfo.Title,
-			URL:    i.prInfo.URL,
-			State:  i.prInfo.State,
-		}
-	}
-
-	return data
-}
-
-// FromInstanceData creates a new Instance from serialized data
-func FromInstanceData(data InstanceData) (*Instance, error) {
-	// Resolve the liveness axis via the shared rollforward (#1108/#1195): prefer
-	// the new `liveness` field, fall back to the legacy `status` int for records
-	// written before #1195, and load a persisted Dead as Lost — recovery-
-	// eligible, which is what makes sessions stranded by an outage under an old
-	// build restorable after an upgrade. A tombstoned record keeps its (Lost)
-	// status; the daemon finishes its teardown rather than restoring it.
-	liveness := livenessFromData(data)
-	// Resolve the in-flight-op axis from the legacy status. A persisted record
-	// is always settled (SaveInstances skips transients), but a SNAPSHOT of a
-	// transient instance carries Loading/Deleting, which buildInstanceFromSnapshot
-	// must reconstruct so the rebuilt row composes to the identical Status.
-	inFlightOp := opForStatus(data.Status)
-	instance := &Instance{
-		ID:           data.ID,
-		Title:        data.Title,
-		Path:         data.Path,
-		Branch:       data.Branch,
-		liveness:     liveness,
-		inFlightOp:   inFlightOp,
-		limitResetAt: data.LimitResetAt,
-		Height:       data.Height,
-		Width:        data.Width,
-		CreatedAt:    data.CreatedAt,
-		UpdatedAt:    data.UpdatedAt,
-		Program:      data.Program,
-		AutoYes:      data.AutoYes,
-		Prompt:       data.Prompt,
-		userKilled:   data.UserKilled,
-		remoteMeta:   data.RemoteMeta,
-	}
-
-	// Pick backend based on persisted BackendType.
-	if data.BackendType == "remote" {
-		hook, err := loadHookBackendForPath(data.Path)
-		if err != nil {
-			return nil, fmt.Errorf("failed to load remote hooks config: %w", err)
-		}
-		instance.backend = hook
-	} else {
-		instance.backend = &LocalBackend{}
-
-		// Preserve backward compatibility: when the branch_created_by_us
-		// field is missing from persisted data (written before this field
-		// was added), default to true. Old saved sessions were created
-		// under the assumption that the session owned the branch, so
-		// keeping that behavior avoids surprising changes on restore.
-		branchCreatedByUs := true
-		if data.Worktree.BranchCreatedByUs != nil {
-			branchCreatedByUs = *data.Worktree.BranchCreatedByUs
-		}
-
-		gw, err := git.NewGitWorktreeFromStorage(
-			data.Worktree.RepoPath,
-			data.Worktree.WorktreePath,
-			data.Worktree.SessionName,
-			data.Worktree.BranchName,
-			data.Worktree.BaseCommitSHA,
-			data.Worktree.ExternalWorktree,
-			branchCreatedByUs,
-		)
-		if err != nil {
-			return nil, fmt.Errorf("failed to restore git worktree: %w", err)
-		}
-		instance.gitWorktree = gw
-
-		// Rebuild the instance's tab list from disk so every tab (agent + shell)
-		// reconnects to its exact tmux session across an af/daemon restart — the
-		// load-bearing #930 requirement. LocalBackend.Start(false) then restores
-		// each tab's session.
-		restoreLocalTabs(instance, data)
-	}
-
-	if data.PRInfo.Number != 0 {
-		instance.prInfo = &git.PRInfo{
-			Number: data.PRInfo.Number,
-			Title:  data.PRInfo.Title,
-			URL:    data.PRInfo.URL,
-			State:  data.PRInfo.State,
-		}
-	}
-
-	// An archived session (#1028) loads INERT: its tmux was torn down and its
-	// worktree moved to the global archive dir at archive time, so there is
-	// nothing to re-spawn or reconnect. Skipping Start leaves started=false and
-	// no tmux binding, which is exactly what makes the status poll (skips
-	// !Started), the Lost-restore loop (gates on ==Lost), and EnsureRootAgents
-	// pass it by — the session sits quiescent until an explicit RestoreArchived.
-	// This is also #970-consistent: a load must never itself un-archive a
-	// session (no worktree move, no spawn) as a side effect. gitWorktree is
-	// already bound above to the persisted (archived) path so restore knows
-	// where the worktree currently lives; the Tabs list restored above is
-	// tmux-less for the same reason (its TmuxName entries reference sessions
-	// that no longer exist, and restoreLocalTabs only binds names, never spawns).
-	if liveness == LiveArchived {
-		return instance, nil
-	}
-
-	if err := instance.Start(false); err != nil {
-		return nil, err
-	}
-
-	return instance, nil
-}
-
-// restoreTmuxSession constructs a tmux session for an exact persisted name. It
-// is a package var (not a direct call) so restore-survival tests can inject
-// mock-backed sessions and stay hermetic; production uses the real constructor.
-var restoreTmuxSession = tmux.NewTmuxSessionFromSanitizedName
-
-// restoreLocalTabs rebuilds a local instance's tab list from persisted data.
-//
-//   - New format (data.Tabs present): each tab is reconstructed in order, and
-//     any tab with a persisted tmux name is bound to that exact session so
-//     LocalBackend.Start can reconnect it across a restart.
-//   - Legacy format (no data.Tabs, written before #930 PR 2): synthesize the
-//     single Agent tab from the legacy TmuxName/Program — keeping the EXACT
-//     legacy tmux name so an existing live agent session survives the upgrade.
-//     No shell tab is synthesized: terminal tabs are on-demand only (#1100).
-func restoreLocalTabs(instance *Instance, data InstanceData) {
-	if len(data.Tabs) > 0 {
-		for idx, td := range data.Tabs {
-			kind := tabKindForData(td.Kind)
-			var ts *tmux.TmuxSession
-			if td.TmuxName != "" {
-				ts = restoreTmuxSession(td.TmuxName, tabProgram(kind, td.Command, data.Program))
-			}
-			var conversation AgentConversationData
-			if td.Conversation != nil {
-				conversation = *td.Conversation
-			} else if idx == 0 && data.AgentConversation != nil {
-				conversation = *data.AgentConversation
-			}
-			instance.Tabs = append(instance.Tabs, &Tab{
-				Name:         td.Name,
-				Kind:         kind,
-				Command:      td.Command,
-				Conversation: conversation,
-				tmux:         ts,
-			})
-		}
-		return
-	}
-
-	// Legacy single-session format: the agent tab keeps its exact legacy name.
-	if data.TmuxName != "" {
-		instance.setTmuxLocked(restoreTmuxSession(data.TmuxName, data.Program))
-	} else {
-		instance.setTmuxLocked(tmux.NewTmuxSession(data.Title, data.Program))
-	}
-	if data.AgentConversation != nil {
-		instance.SetAgentConversation(*data.AgentConversation)
-	}
-}
-
 // tabProgram resolves the program a tab's tmux session runs, by kind: the agent
 // program for Agent tabs, $SHELL for Shell tabs, and the explicit command for
 // Process tabs (falling back to $SHELL when empty).
@@ -839,537 +576,4 @@ func defaultShell() string {
 		shell = "/bin/sh"
 	}
 	return shell
-}
-
-// Options for creating a new instance
-type InstanceOptions struct {
-	// Title is the title of the instance.
-	Title string
-	// Path is the path to the workspace.
-	Path string
-	// Program is the program to run in the instance (e.g. "claude", "aider --model ollama_chat/gemma3:1b")
-	Program string
-	// If AutoYes is true, then
-	AutoYes bool
-	// ForceRemote forces the instance to use the remote hook backend,
-	// even if the repo config would default to local.
-	ForceRemote bool
-	// InPlace attaches the session to the repo's existing working tree at its
-	// current branch (`af sessions create --here`) instead of creating a new
-	// git worktree+branch. The worktree is marked external so kill/cleanup
-	// never removes the user's tree or branch. Local backend only.
-	InPlace bool
-}
-
-// backendFactory constructs the Backend used by a new Instance. It is a
-// package-level variable (not a hard-coded branch) so tests can inject a
-// FakeBackend through SetBackendFactoryForTest without touching production
-// code paths. Defaults to the real local/remote branching.
-var backendFactory = defaultBackendFactory
-
-func defaultBackendFactory(opts InstanceOptions, absPath string) (Backend, error) {
-	if opts.ForceRemote {
-		hook, err := loadHookBackendForPath(absPath)
-		if err != nil {
-			return nil, fmt.Errorf("remote hooks not configured for this repo: %w", err)
-		}
-		return hook, nil
-	}
-	return &LocalBackend{}, nil
-}
-
-// SetBackendFactoryForTest replaces the backend factory with f and returns a
-// restore function. Intended for use in tests that need to swap in a
-// FakeBackend so NewInstance-driven creation flows stay on the hot path.
-func SetBackendFactoryForTest(f func(opts InstanceOptions, absPath string) (Backend, error)) func() {
-	prev := backendFactory
-	backendFactory = f
-	return func() { backendFactory = prev }
-}
-
-// newSessionID mints a random RFC-4122 v4 UUID for an instance's stable identity
-// (#1195). It is a package var so tests can inject deterministic IDs. crypto/rand
-// is the entropy source; on the (near-impossible) read failure it falls back to a
-// timestamp-derived value so session creation never blocks on entropy — still
-// unique per call in practice, and the reconcile's title+CreatedAt fallback covers
-// any theoretical collision.
-var newSessionID = func() string {
-	var b [16]byte
-	if _, err := rand.Read(b[:]); err != nil {
-		return fmt.Sprintf("ts-%d", time.Now().UnixNano())
-	}
-	b[6] = (b[6] & 0x0f) | 0x40 // version 4
-	b[8] = (b[8] & 0x3f) | 0x80 // variant 10
-	return fmt.Sprintf("%x-%x-%x-%x-%x", b[0:4], b[4:6], b[6:8], b[8:10], b[10:16])
-}
-
-func NewInstance(opts InstanceOptions) (*Instance, error) {
-	t := time.Now()
-
-	// An in-place session runs in the repo's local working tree; a remote
-	// session has no local worktree at all — the two are contradictory.
-	if opts.InPlace && opts.ForceRemote {
-		return nil, fmt.Errorf("remote sessions cannot run in-place in the local repo working tree")
-	}
-
-	// Convert path to absolute
-	absPath, err := filepath.Abs(opts.Path)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get absolute path: %w", err)
-	}
-
-	backend, err := backendFactory(opts, absPath)
-	if err != nil {
-		return nil, err
-	}
-
-	return &Instance{
-		ID:        newSessionID(),
-		Title:     opts.Title,
-		liveness:  LiveReady,
-		Path:      absPath,
-		Program:   opts.Program,
-		Height:    0,
-		Width:     0,
-		CreatedAt: t,
-		UpdatedAt: t,
-		AutoYes:   opts.AutoYes,
-		inPlace:   opts.InPlace,
-		backend:   backend,
-	}, nil
-}
-
-func (i *Instance) RepoName() (string, error) {
-	if i.IsRemote() {
-		return "", fmt.Errorf("remote instances do not have a local repo")
-	}
-	i.mu.RLock()
-	started := i.started
-	gw := i.gitWorktree
-	i.mu.RUnlock()
-	if !started {
-		return "", fmt.Errorf("cannot get repo name for instance that has not been started")
-	}
-	if gw == nil {
-		return "", fmt.Errorf("cannot get repo name for instance without a git worktree")
-	}
-	return gw.GetRepoName(), nil
-}
-
-// SetAutoYes sets the AutoYes flag under the instance mutex. Writers must use
-// this rather than assigning i.AutoYes directly: TapEnter runs from the
-// metadata-tick background goroutine and reads AutoYes under i.mu.RLock, so
-// any unsynchronized write produces a data race (issue #563, regression from
-// PR #560 which moved the tick off the bubbletea event loop).
-func (i *Instance) SetAutoYes(autoYes bool) {
-	i.mu.Lock()
-	defer i.mu.Unlock()
-	i.AutoYes = autoYes
-}
-
-// GetBranch returns the current worktree branch name under the Instance's
-// mutex. Readers that run from goroutines other than the one mutating the
-// instance (notably the bubbletea renderer) must use this accessor rather
-// than reading i.Branch directly, or the race detector flags a write in
-// LocalBackend.Start vs a read in InstanceRenderer.Render.
-func (i *Instance) GetBranch() string {
-	i.mu.RLock()
-	defer i.mu.RUnlock()
-	return i.Branch
-}
-
-// MarkUserKilled records kill intent on the instance (#1108). Callers persist
-// the instance afterwards so the tombstone survives a daemon crash mid-kill.
-func (i *Instance) MarkUserKilled() {
-	i.mu.Lock()
-	defer i.mu.Unlock()
-	i.userKilled = true
-}
-
-// UserKilled reports whether an explicit kill was recorded for this instance.
-func (i *Instance) UserKilled() bool {
-	i.mu.RLock()
-	defer i.mu.RUnlock()
-	return i.userKilled
-}
-
-// firstTimeSetup is true if this is a new instance. Otherwise, it's one loaded from storage.
-func (i *Instance) Start(firstTimeSetup bool) error {
-	return i.backend.Start(i, firstTimeSetup)
-}
-
-// Kill terminates the instance and cleans up all resources
-func (i *Instance) Kill() error {
-	return i.backend.Kill(i)
-}
-
-// Recover re-establishes a Lost instance's backing session (#1108). Called by
-// the daemon's restore loop and by user-initiated restore (#1300); loads stay
-// side-effect free (#970).
-func (i *Instance) Recover() error {
-	return i.backend.Recover(i)
-}
-
-// Respawn re-establishes the instance's backing session in place without a
-// liveness precondition — the guard-free core of Recover. The usage-limit
-// manual-retry (#1146, resumeFromLimit) uses it to re-spawn an agent that exited
-// while blocked at a limit wall: that session is LiveLimitReached, which Recover's
-// !Lost guard rejects, but the re-spawn mechanics are identical. The caller owns
-// the precondition.
-func (i *Instance) Respawn() error {
-	return i.backend.Respawn(i)
-}
-
-// ArchiveTeardown tears down every tab's tmux session for an archive AND
-// relocates the worktree to dest in one operation (#1028) — the tmux half of
-// Kill, but it PRESERVES the record and MOVES the worktree instead of deleting
-// it. It routes through the shared teardownTabs core in the archive mode, so the
-// #802 "wait for every pane to exit before touching the worktree" ordering is
-// shared code with Kill rather than the duplicated prose it was when the move
-// lived in a separate daemon step (#1195 Phase 2b). It is deliberately
-// best-effort for tmux (a stuck session only logs, mirroring Kill) and:
-//   - keeps the AGENT tab's tmux binding (its session name) so a failed archive
-//     can re-spawn it in place via the Lost-restore loop;
-//   - drops the shell/process tabs entirely — only the agent session is brought
-//     back on un-archive (Sachin's #1028 requirement);
-//   - leaves gitWorktree and started untouched, so the daemon caller controls
-//     the final state (started=false + Archived on success; Lost on a failed
-//     move — returned here — where started stays true so the loop re-spawns the
-//     agent).
-//
-// Returns the worktree-move error (nil on success). Local instances only —
-// remote sessions have no local tmux/worktree and the daemon rejects archiving
-// them before reaching here.
-func (i *Instance) ArchiveTeardown(dest string) error {
-	return i.teardownTabs(teardownArchive{dest: dest})
-}
-
-// SetArchived flips the instance into the inert Archived state atomically:
-// started=false (no tmux binding backs it) and liveness=Archived, clearing any
-// in-flight op. Called by the daemon after a successful archive move.
-func (i *Instance) SetArchived() {
-	i.mu.Lock()
-	defer i.mu.Unlock()
-	i.started = false
-	i.liveness = LiveArchived
-	i.inFlightOp = OpNone
-}
-
-// RestoreArchivedWorktree moves this instance's archived worktree back to dest
-// and re-registers it against the origin repo (#1028). Surfaces git.ErrRepoGone
-// when the repo has been deleted so the caller can leave the archive intact.
-func (i *Instance) RestoreArchivedWorktree(dest string) error {
-	i.mu.RLock()
-	gw := i.gitWorktree
-	i.mu.RUnlock()
-	if gw == nil {
-		return fmt.Errorf("cannot restore %q: instance has no worktree", i.Title)
-	}
-	return gw.RestoreWorktreeTo(dest)
-}
-
-// RestoreFromArchive re-spawns an archived instance's agent after its worktree
-// has been moved back into place (#1028), flipping it live. It marks the
-// instance started + Lost so the Recover re-spawn path is eligible (the same
-// re-spawn the #1108 Lost-restore loop drives), then Recover brings the agent
-// session up and sets Running (markLive clears the OpRestoring fence). On a
-// Recover failure the instance is dropped to a plain Lost (op cleared), so the
-// daemon's Lost-restore loop keeps retrying — the worktree is already back in
-// place, so the session self-heals rather than stranding as Archived with no
-// tmux. Only the agent tab is restored (shell/process tabs were dropped at
-// archive time, per #1028).
-//
-// liveness is set to Lost (so Recover's ==Lost gate accepts it) and OpRestoring
-// fences the re-spawn window: the daemon poll skips an instance with an
-// in-flight op, so it never probes the half-spawned session and marks it Lost
-// out from under the restore. This replaces the old "park it in Lost purely to
-// trigger the re-spawn loop" overload (#1195).
-func (i *Instance) RestoreFromArchive() error {
-	// Enter the restore fence through the chokepoint (#1195 Phase 2d): BeginRestore
-	// is legal only from Archived and sets started=true + Lost + OpRestoring — the
-	// exact head this used to write by hand, now enforcing I3 (a restore may begin
-	// only from an archived session; no double-restore).
-	if err := i.Transition(BeginRestore()); err != nil {
-		return err
-	}
-	if err := i.backend.Recover(i); err != nil {
-		// Re-spawn failed: drop the fence to a plain Lost (started left true) so
-		// the #1108 restore loop owns the retry against the now-restored worktree.
-		_ = i.Transition(AbortRestoreToLost())
-		return err
-	}
-	return nil
-}
-
-// CloseAttachOnly releases the resources this instance opened to view or drive
-// its session (a tmux attach PTY, a remote preview process) without destroying
-// the session, worktree, or remote record. Use it — never Kill — to discard a
-// duplicate Instance built from disk that lost a race to the canonical tracked
-// Instance (#867); see Backend.CloseAttachOnly.
-func (i *Instance) CloseAttachOnly() error {
-	return i.backend.CloseAttachOnly(i)
-}
-
-func (i *Instance) Preview() (string, error) {
-	return i.backend.Preview(i)
-}
-
-func (i *Instance) HasUpdated() (updated bool, hasPrompt bool, content string) {
-	return i.backend.HasUpdated(i)
-}
-
-// CheckAndHandleTrustPrompt checks for and dismisses the trust prompt for supported programs.
-func (i *Instance) CheckAndHandleTrustPrompt() bool {
-	return i.backend.CheckAndHandleTrustPrompt(i)
-}
-
-// TapEnter sends an enter key press to the tmux session if AutoYes is enabled.
-func (i *Instance) TapEnter() {
-	i.backend.TapEnter(i)
-}
-
-func (i *Instance) Attach() (chan struct{}, error) {
-	return i.backend.Attach(i)
-}
-
-func (i *Instance) SetPreviewSize(width, height int) error {
-	return i.backend.SetPreviewSize(i, width, height)
-}
-
-// GetGitWorktree returns the git worktree for the instance
-func (i *Instance) GetGitWorktree() (*git.GitWorktree, error) {
-	i.mu.RLock()
-	defer i.mu.RUnlock()
-	if !i.started {
-		return nil, fmt.Errorf("cannot get git worktree for instance that has not been started")
-	}
-	return i.gitWorktree, nil
-}
-
-// GetWorktreePath returns the worktree path for the instance, or empty string if unavailable
-func (i *Instance) GetWorktreePath() string {
-	i.mu.RLock()
-	gw := i.gitWorktree
-	i.mu.RUnlock()
-
-	if gw == nil {
-		return ""
-	}
-	return gw.GetWorktreePath()
-}
-
-// GetRepoPath returns the resolved git repo path stored in the instance's
-// worktree, or empty string when no worktree is attached (e.g. a remote-
-// backend instance). Callers using the result to derive a repo ID must
-// fall back to Instance.Path when this is empty (#667).
-func (i *Instance) GetRepoPath() string {
-	i.mu.RLock()
-	gw := i.gitWorktree
-	i.mu.RUnlock()
-
-	if gw == nil {
-		return ""
-	}
-	return gw.GetRepoPath()
-}
-
-func (i *Instance) Started() bool {
-	i.mu.RLock()
-	defer i.mu.RUnlock()
-	return i.started
-}
-
-// IsExternalWorktree reports whether the instance's worktree is external/in-place
-// (`af sessions create --here`, or a legacy external record) — the same flag
-// MoveWorktree checks. Such a worktree is the user's own working tree and must
-// never be relocated, so the daemon rejects archiving it (#1028). Returns false
-// when the instance has no worktree yet.
-func (i *Instance) IsExternalWorktree() bool {
-	i.mu.RLock()
-	gw := i.gitWorktree
-	i.mu.RUnlock()
-	return gw != nil && gw.IsExternalWorktree()
-}
-
-// SetTitle sets the title of the instance. Returns an error if the instance has started.
-// We cant change the title once it's been used for a tmux session etc.
-func (i *Instance) SetTitle(title string) error {
-	i.mu.Lock()
-	defer i.mu.Unlock()
-	if i.started {
-		return fmt.Errorf("cannot change title of a started instance")
-	}
-	i.Title = title
-	return nil
-}
-
-// TmuxAlive returns true if the underlying session is alive.
-// For remote backends this delegates to IsAlive.
-func (i *Instance) TmuxAlive() bool {
-	return i.backend.IsAlive(i)
-}
-
-// ResolvedAgent returns the canonical agent (one of tmux.SupportedPrograms)
-// this instance's pane will actually run, or "" when the resolved command
-// runs no known agent — e.g. a program_overrides entry pointing an agent name
-// at a plain shell (#1131). Agent-specific behavior (readiness heuristics,
-// trust-prompt handling, flag injection) must key off this, never off
-// Instance.Program: Program is the config-name enum the instance was created
-// with, and an override may point it at a different program entirely (#1116).
-//
-// Once the tmux session exists, its program string (override-resolved and
-// flag-injected by Start) is the ground truth. Before Start — or in tests
-// that never attach a tmux session — detection falls back to the raw Program
-// value, which also covers legacy free-form persisted values like
-// "/home/foo/bin/claude --plugin-dir x" (#677).
-func (i *Instance) ResolvedAgent() string {
-	i.mu.RLock()
-	ts := i.tmuxLocked()
-	i.mu.RUnlock()
-	if ts != nil {
-		if p := ts.Program(); strings.TrimSpace(p) != "" {
-			return tmux.DetectAgentFromCommand(p)
-		}
-	}
-	return tmux.DetectAgentFromCommand(i.Program)
-}
-
-// GetPRInfo returns the associated GitHub PR info, or nil if none.
-func (i *Instance) GetPRInfo() *git.PRInfo {
-	i.mu.RLock()
-	defer i.mu.RUnlock()
-	return i.prInfo
-}
-
-// SetPRInfo sets the associated GitHub PR info.
-func (i *Instance) SetPRInfo(info *git.PRInfo) {
-	i.mu.Lock()
-	defer i.mu.Unlock()
-	i.prInfo = info
-	i.prInfoLastFetched = time.Now()
-}
-
-// PRInfoAge returns how long ago PR info was last fetched. Returns a very
-// large duration if PR info has never been fetched in this process.
-func (i *Instance) PRInfoAge() time.Duration {
-	i.mu.RLock()
-	defer i.mu.RUnlock()
-	if i.prInfoLastFetched.IsZero() {
-		return time.Duration(1<<62 - 1)
-	}
-	return time.Since(i.prInfoLastFetched)
-}
-
-// MarkPRInfoFetched bumps the fetch timestamp without touching the cached
-// value. Used after a transient fetch error so we don't re-try on every
-// subsequent selection change.
-func (i *Instance) MarkPRInfoFetched() {
-	i.mu.Lock()
-	defer i.mu.Unlock()
-	i.prInfoLastFetched = time.Now()
-}
-
-// FetchPRInfoSnapshot returns the data needed to fetch PR info for this
-// instance off the main event loop. The returned repoPath is empty when the
-// instance is not ready for fetching (not started, no worktree, or remote).
-func (i *Instance) FetchPRInfoSnapshot() (repoPath, branch string) {
-	i.mu.RLock()
-	defer i.mu.RUnlock()
-	if !i.started || i.gitWorktree == nil {
-		return "", ""
-	}
-	return i.gitWorktree.GetRepoPath(), i.Branch
-}
-
-// SendPrompt sends a prompt to the session
-func (i *Instance) SendPrompt(prompt string) error {
-	return i.backend.SendPrompt(i, prompt)
-}
-
-// SendPromptCommand sends a prompt using a more reliable command-based approach.
-// This is more reliable for headless/scheduled runs where the PTY may not persist.
-func (i *Instance) SendPromptCommand(prompt string) error {
-	return i.backend.SendPromptCommand(i, prompt)
-}
-
-// PreviewFullHistory captures the entire session output including full scrollback history
-func (i *Instance) PreviewFullHistory() (string, error) {
-	return i.backend.PreviewFullHistory(i)
-}
-
-// SetTmuxSession sets the agent tab's tmux session for testing purposes,
-// materializing the single Agent tab if needed.
-func (i *Instance) SetTmuxSession(session *tmux.TmuxSession) {
-	i.setTmuxLocked(session)
-}
-
-// SetStartedForTest toggles the started flag for testing purposes. Prefer
-// Start() in non-test code; this exists so unit tests can exercise flows
-// gated on Started() without spinning up a real tmux session.
-func (i *Instance) SetStartedForTest(started bool) {
-	i.mu.Lock()
-	defer i.mu.Unlock()
-	i.started = started
-}
-
-// SetGitWorktreeForTest assigns a git worktree to this instance. Test-only:
-// the real flow sets this inside LocalBackend.Start, which isn't available
-// in unit tests that use FakeBackend.
-func (i *Instance) SetGitWorktreeForTest(gw *git.GitWorktree) {
-	i.mu.Lock()
-	defer i.mu.Unlock()
-	i.gitWorktree = gw
-}
-
-// AddTabForTest appends a tmux-less tab record. Test-only: UI tests (the
-// sidebar tree, tab labels) need instances with a populated tab LIST without
-// spinning up real tmux sessions; the tab is never attachable or previewable.
-func (i *Instance) AddTabForTest(name string, kind TabKind) {
-	i.mu.Lock()
-	defer i.mu.Unlock()
-	i.Tabs = append(i.Tabs, &Tab{Name: name, Kind: kind})
-}
-
-// SendKeys sends keys to the underlying session. For remote backends this
-// returns an explicit error since raw key injection is not supported.
-func (i *Instance) SendKeys(keys string) error {
-	return i.backend.SendKeys(i, keys)
-}
-
-// IsRemote returns true if this instance uses the remote hook backend.
-func (i *Instance) IsRemote() bool {
-	if i.backend == nil {
-		return false
-	}
-	return i.backend.Type() == "remote"
-}
-
-// SupportsRemoteTerminal reports whether this instance can open an interactive
-// terminal on its remote machine — i.e. it uses the remote hook backend and
-// the optional terminal_cmd hook is configured (#843).
-func (i *Instance) SupportsRemoteTerminal() bool {
-	hb, ok := i.backend.(*HookBackend)
-	return ok && hb.HasTerminalCmd()
-}
-
-// AttachRemoteTerminal opens an interactive terminal on the remote machine via
-// the terminal_cmd hook. The returned channel is closed when the user detaches
-// or the terminal_cmd process exits. Errors when the instance is not backed by
-// remote hooks or terminal_cmd is not configured.
-func (i *Instance) AttachRemoteTerminal() (chan struct{}, error) {
-	hb, ok := i.backend.(*HookBackend)
-	if !ok {
-		return nil, fmt.Errorf("remote terminal is only available for remote sessions")
-	}
-	return hb.AttachTerminal(i)
-}
-
-// GetBackend returns the backend for the instance (mainly for testing).
-func (i *Instance) GetBackend() Backend {
-	return i.backend
-}
-
-// SetBackend sets the backend for the instance (mainly for testing).
-func (i *Instance) SetBackend(b Backend) {
-	i.backend = b
 }
